Extract initialize() to test-utils module
So the initialize method can be used in other test files, too.
